refactor (Lotto) : BonusNumber를 리팩토링 한다
- 정적 팩토리 메서드를 사용
- lotto 패키지로 분리
