+++ src/test/java/lotto/domain/BonusNumberTest.java
@@ -2,6 +2,7 @@ package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import lotto.domain.lotto.BonusNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -11,10 +12,10 @@ class BonusNumberTest {
     @DisplayName("번호가 1 미만 45 초과라면 예외를 발생한다")
     @Test
     void createNumberByOutOfRange() {
-        assertThatThrownBy(() -> new BonusNumber(0))
+        assertThatThrownBy(() -> BonusNumber.from(0))
             .isInstanceOf(IllegalArgumentException.class);
 
-        assertThatThrownBy(() -> new BonusNumber(46))
+        assertThatThrownBy(() -> BonusNumber.from(46))
             .isInstanceOf(IllegalArgumentException.class);
     }
 }
